backend/requirements.txt
appdirs==1.4.4
asgiref==3.8.1
baron==0.10.1
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.1
django-cache-url==3.4.5
django-cors-headers==4.4.0
django-environs==0.1.1
django-phonenumber-field==8.0.0
django-user-sessions==2.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environs==11.0.0
marshmallow==3.21.3
packaging==24.1
phonenumbers==8.13.42
pillow==10.4.0
PyJWT==2.9.0
python-dotenv==1.0.1
redbaron==0.9.2
rply==0.7.8
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
